backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pymongo==4.12.0
python-dotenv==1.1.0
